created classes for keyshare and secretshare. Added word doc
