+++ network_algs/__init__.py
@@ -1,1 +1,4 @@
-from .alt_path_primitive import *
+# from .alt_path_primitive import *
+from .network_base import *
+from .ShareKey import *
+from .ShareSecret import *
